feat: run telemetry collection as a background process
